internal: move {config,db,http} into internal
