+++ internal/db/repl.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/peterh/liner"
 
-	"github.com/google/cayley/config"
 	"github.com/google/cayley/graph"
+	"github.com/google/cayley/internal/config"
 	"github.com/google/cayley/quad/cquads"
 	"github.com/google/cayley/query"
 	"github.com/google/cayley/query/gremlin"
